fix: restore clean migration runner and add rate_limit_cooldown_hits table to migrations
